skeleton decomposition added

+++ ppm/ppm.cpp
@@ -3,7 +3,7 @@
 using namespace std;
 
 
-int create_ppm(int N, int M, double const * const data, std::string filename, double maxi) {
+int create_ppm(int N, int M, float const * const data, std::string filename, double maxi) {
     int img_N = 256;
     int img_M = 256;
     size_t SIZE = img_N * img_M;
