+++ ppm/ppm.h
@@ -4,4 +4,4 @@
 #include <cmath>
 #include <omp.h>
 
-int create_ppm(int N, int M, double const * const data, std::string filename, double maxi);
+int create_ppm(int N, int M, float const * const data, std::string filename, double maxi);
